Add files via upload
RC Robot Vacuum. Made out of a broken Wyze robot vacuum. Added 2 775 DC motors with 2 CL9030 DC motor controllers. The motors are paired to a 3D printed 3:1 gearbox, making for a theoretical top speed of 23 mph. Everything is mounted on a 3d printed base, which is bolted to the frame. Everything in this build is custom-made by Prithvi.